requirements.txt
yfinance
pandas
requests
beautifulsoup4
google-genai